[REFACTOR] UIComponent분리 (#17)
