+++ Manito/Manito/Screens/Detail-Wait/DetailWaitViewController.swift
@@ -15,12 +15,6 @@ class DetailWaitViewController: BaseViewController {
     var maxUser = 15
     lazy var userCount = userArr.count
 
-    private enum StartStatus: String {
-        case waiting = "대기중"
-        case starting = "진행중"
-        case complete = "완료"
-    }
-
     private enum ButtonText: String {
         case waiting = "시작을 기다리는 중..."
         case start = "마니또 시작"
@@ -37,48 +31,7 @@ class DetailWaitViewController: BaseViewController {
         return button
     }()
 
-    private let roomTitle: UILabel = {
-        let label = UILabel()
-        label.text = "명예소방관"
-        label.textColor = .white
-        label.font = .font(.regular, ofSize: 34)
-        return label
-    }()
-
-    private let startStauts: UILabel = {
-        let label = UILabel()
-        label.text = StartStatus.waiting.rawValue
-        label.backgroundColor = .badgeBeige
-        label.layer.masksToBounds = true
-        label.layer.cornerRadius = 11
-        label.textColor = .darkGrey004
-        label.font = .font(.regular, ofSize: 13)
-        label.textAlignment = .center
-        return label
-    }()
-
-    private lazy var durationView: UIView = {
-        let durationView = UIView()
-        durationView.backgroundColor = .darkRed.withAlphaComponent(0.65)
-        durationView.layer.cornerRadius = 8
-        return durationView
-    }()
-
-    private let durationText: UILabel = {
-        let durationText = UILabel()
-        durationText.text = "진행 기간"
-        durationText.textColor = .grey004
-        durationText.font = .font(.regular, ofSize: 14)
-        return durationText
-    }()
-
-    private let durationDateText: UILabel = {
-        let dateText = UILabel()
-        dateText.text = "22.06.06 ~ 22.06.10"
-        dateText.textColor = .white
-        dateText.font = .font(.regular, ofSize: 18)
-        return dateText
-    }()
+    private let titleView = DetailWaitTitleView()
 
     private let togetherFriendText: UILabel = {
         let label = UILabel()
@@ -134,44 +87,17 @@ class DetailWaitViewController: BaseViewController {
     }
 
     override func render() {
-        view.addSubview(roomTitle)
-        roomTitle.snp.makeConstraints {
-            $0.leading.equalToSuperview().inset(16)
-            $0.top.equalTo(view.safeAreaLayoutGuide).inset(30)
-        }
-
-        view.addSubview(startStauts)
-        startStauts.snp.makeConstraints {
-            $0.centerY.equalTo(roomTitle.snp.centerY)
-            $0.leading.equalTo(roomTitle.snp.trailing).offset(10)
-            $0.width.equalTo(66)
-            $0.height.equalTo(23)
-        }
-
-        durationView.addSubview(durationText)
-        durationView.addSubview(durationDateText)
-
-        view.addSubview(durationView)
-        durationView.snp.makeConstraints {
-            $0.top.equalTo(roomTitle.snp.bottom).offset(30)
-            $0.trailing.leading.equalToSuperview().inset(16)
-            $0.height.equalTo(36)
-        }
-
-        durationText.snp.makeConstraints {
-            $0.leading.equalToSuperview().inset(20)
-            $0.centerY.equalToSuperview()
-        }
-
-        durationDateText.snp.makeConstraints {
-            $0.trailing.equalToSuperview().inset(40)
-            $0.centerY.equalToSuperview()
+        view.addSubview(titleView)
+        titleView.snp.makeConstraints {
+            $0.leading.trailing.equalToSuperview().inset(16)
+            $0.top.equalToSuperview().offset(100)
+            $0.height.equalTo(86)
         }
 
         view.addSubview(togetherFriendText)
         togetherFriendText.snp.makeConstraints {
             $0.leading.equalToSuperview().inset(16)
-            $0.top.equalTo(durationView.snp.bottom).offset(44)
+            $0.top.equalTo(titleView.snp.bottom).offset(44)
         }
 
         view.addSubview(comeInText)
